[Glitch] Split VerifiedBadge from Account component
Partial port of 0461f83320378fb8cee679da896ce35cec5bcbf3 to glitch-soc
